@@ -1,12 +1,12 @@
 {
   "name": "twisted",
-  "version": "1.53.0",
+  "version": "1.55.0",
   "description": "Fetching riot games api data",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "lint:fix": "tslint --project tsconfig.json --fix",
+    "lint": "eslint ./src/**/*",
+    "lint:fix": "eslint --fix ./src/**/*",
     "test": "jest",
     "test:coverage": "jest --coverage",
     "build": "tsc",
@@ -41,31 +41,30 @@
   },
   "license": "ISC",
   "devDependencies": {
-    "@types/chai": "^4.2.1",
+    "@types/chai": "^4.3.11",
     "@types/dotenv": "^6.1.1",
-    "@types/http-status-codes": "^1.2.0",
-    "@types/jest": "^25.2.1",
-    "@types/lodash": "^4.14.138",
-    "@types/node": "^14.6.4",
-    "@types/promise-queue": "^2.2.0",
-    "@types/sinon": "^7.0.13",
-    "@typescript-eslint/eslint-plugin": "^2.1.0",
-    "@typescript-eslint/parser": "^2.1.0",
-    "jest": "^25.4.0",
-    "ts-jest": "^25.4.0",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.19.0",
-    "tslint-config-standard": "^8.0.1",
-    "typescript": "^3.8.3"
+    "@types/jest": "^29.5.10",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.10.0",
+    "@types/promise-queue": "^2.2.3",
+    "@types/sinon": "^17.0.2",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "^6.12.0",
+    "eslint": "^8.54.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jsdoc": "^46.9.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.3.2"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "camelcasejs": "^1.0.3",
-    "dotenv": "^8.1.0",
-    "http-status-codes": "^1.3.2",
-    "lodash": "^4.17.15",
+    "axios": "^1.6.2",
+    "dotenv": "^16.3.1",
+    "http-status-codes": "^2.3.0",
+    "lodash": "^4.17.21",
     "promise-queue": "^2.2.5",
-    "uuid": "^8.3.0"
+    "uuid": "^9.0.1"
   },
   "standard": {
     "env": [
